Inverted ifs when making move. Added function that checks if board is full, so program can call a draw.  Cleaned up penteEngine, by wrapping first minmax call in function. So all evaluation AIs can use it. Formatted settings menu and changed many cin to _getch, so input is faster

+++ pente/source/graphicalMode.cpp
@@ -10,8 +10,8 @@ void mouse_button_callback(GLFWwindow* window, int button, int action, int mods)
 	{
 		double xpos, ypos;
 		glfwGetCursorPos(window, &xpos, &ypos);
-		communication.x = round((xpos - 60) / disBetweenSquares);
-		communication.y = round((ypos - 60) / disBetweenSquares);
+		communication.x = (int)round((xpos - 60) / disBetweenSquares);
+		communication.y = (int)round((ypos - 60) / disBetweenSquares);
 	}
 }
